Fixing incorrect order of items
Things were being referenced before they had been defined.  Strangely, this had not previously caused problems before the move to odoo.sh.  It was not a problem on the production branch which was an import from community (so gg2, our custom module was never completely re-installed and tested as it is in the dev branch)